+++ src/icons/EazeStormIcons.java
@@ -1,0 +1,11 @@
+package icons;
+
+import com.intellij.icons.AllIcons;
+
+import javax.swing.*;
+
+public class EazeStormIcons {
+    public static final Icon QuickFix = AllIcons.Actions.QuickfixBulb;
+    public static final Icon Edit = AllIcons.Actions.EditSource;
+    public static final Icon Export = AllIcons.Actions.Export;
+}
